Bugfix: broker Channel not registering pool used to create new Client
The kombu-redis Channel class stores the first created pool in
self._pool. What was actually needed is an overload of the Channel
_get_pool method. Not doing so leads to weird bugs down the road with
redis-py trying to release apparently unregistered connections. Took me
a while to nail this one!
The duplicated code from the Channel _connparams method was removed as
it's not needed.

If someone reads this, it would be nice to clean this code to have it
merged upstream in celery, cf. issue #3.